% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{contact_countries}
\alias{contact_countries}
\title{Mapping of ESEN2 countries to POLYIMOD countries}
\description{
Mapping of ESEN2 countries to POLYIMOD countries
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
