% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{esen2_adjImm}
\alias{esen2_adjImm}
\title{Contact-adjusted immunity levels derived from ESEN2 serology}
\description{
Contact-adjusted immunity levels derived from ESEN2 serology
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
