% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{esen2_schedules}
\alias{esen2_schedules}
\title{Vaccination schedules in ESEN2 countries as per Andrews et al. (2008)}
\description{
Vaccination schedules in ESEN2 countries as per Andrews et al. (2008)
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
