% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{mcv_world_estimate}
\alias{mcv_world_estimate}
\title{WUENIC estimates of measles vaccination, downloaded from the World Health Organization: https://www.who.int/immunization/monitoring_surveillance/data/en/}
\description{
WUENIC estimates of measles vaccination, downloaded from the World Health Organization: https://www.who.int/immunization/monitoring_surveillance/data/en/
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
