% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{ms_world}
\alias{ms_world}
\title{Worldwide measles case data, downloaded from the World Health Organization: https://www.who.int/immunization/monitoring_surveillance/data/en/}
\description{
Worldwide measles case data, downloaded from the World Health Organization: https://www.who.int/immunization/monitoring_surveillance/data/en/
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
