% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{scenarios_adjImm}
\alias{scenarios_adjImm}
\title{Contact-adjusted immunity levels in immunity scenarios}
\description{
Contact-adjusted immunity levels in immunity scenarios
}
\author{
Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}
