% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_sero_ds.r
\name{toy_sero_ds}
\alias{toy_sero_ds}
\title{Create a toy individualised data set of seroprevalence}
\usage{
toy_sero_ds(n, pos_eqi, survey.yr, lower.age.limits)
}
\arguments{
\item{n}{number of sample to generate per country}

\item{pos_eqi}{a named list of vectors of size 3, each name corresponding to the country represented, and the three elements of the vector corresponding to the proportion positive and equivocal (the reminder being negative)}

\item{survey.yr}{an integer vector: year(s) during which the seroprevalence studies occurred}

\item{lower.age.limits}{an integer vector: lower limits of the age groups for which to generate data}
}
\value{
a data frame representing a toy data set of individual serology results, with columns \code{survey.yr} (the year of the survey), \code{country} (the country), \code{age1} (the lower limit of the age group) and \code{stdres} (the standardised result)
}
\description{
Create a toy individualised data set of seroprevalence
}
\examples{
  ms.sero <- toy_sero_ds(survey.yr=2002, pos_eqi=list(Spain=c(0.8, 0.1), Ireland=c(0.9, 0)), lower.age.limits=seq(0, 70, by=5), n=100)
}
\author{
Sebastian Funk <sebastian.funk@lshtm.ac.uk>
}
